+++ iamf/cli/user_metadata_builder/codec_config_obu_metadata_builder.cc
@@ -28,8 +28,6 @@ CodecConfigObuMetadataBuilder::GetLpcmCodecConfigObuMetadata(
   CodecConfigObuMetadata codec_config_obu_metadata;
   codec_config_obu_metadata.set_codec_config_id(codec_config_id);
   auto* codec_config = codec_config_obu_metadata.mutable_codec_config();
-  // Set codec id as ipcm.
-  codec_config->set_codec_id(iamf_tools_cli_proto::CODEC_ID_LPCM);
   codec_config->set_num_samples_per_frame(num_samples_per_frame);
 
   auto* decode_config = codec_config->mutable_decoder_config_lpcm();
@@ -55,7 +53,6 @@ CodecConfigObuMetadataBuilder::GetOpusCodecConfigObuMetadata(
   codec_config_obu_metadata.set_codec_config_id(codec_config_id);
   auto& codec_config = *codec_config_obu_metadata.mutable_codec_config();
 
-  codec_config.set_codec_id(iamf_tools_cli_proto::CODEC_ID_OPUS);
   codec_config.set_num_samples_per_frame(num_samples_per_frame);
   codec_config.mutable_decoder_config_opus()->set_version(kDefaultOpusVersion);
   codec_config.mutable_decoder_config_opus()->set_input_sample_rate(